Extracted XML persistence functionality
